add igloo music xml generator
