+++ src/server/routes/game.ts
@@ -5,7 +5,6 @@ import { HttpServer } from "../http";
 import { SettingsManager } from "../settings";
 import { getSetupTxt } from "./setup.txt";
 import { getNewsTxt } from "./news.txt";
-import { getDynamicMusicListData } from "../timelines/igloo-lists";
 import { isEngine2, isEngine3 } from "../timelines/dates";
 import { FRAME_HACKS } from '@server/game-data/frame-hacks';
 import { fromLE, parseSwf } from "@common/flash/parser";
@@ -96,11 +95,6 @@ export function createHttpServer(settingsManager: SettingsManager): HttpServer {
   );
 
   // text file generating
-
-  // serving dynamic igloo data for ben/randomno's dynamic igloo music list mod
-  server.getData('play/v2/content/global/en/igloo_music.xml', (s) => {
-    return getDynamicMusicListData(s.settings.version);
-  });
 
   server.getData('load.swf', (s) => {
     const file = s.settings.fps30 ? 'tool:load30.swf' : 'tool:load.swf';
